requirement.txt
amqp==5.0.2
android==1.0
asgiref==3.3.1
asn1crypto==1.4.0
astroid==2.4.2
attrs==20.3.0
audiostream==0.2
Babel==2.9.0
base58==2.0.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bitarray==1.2.2
bitcash==0.6.0
bitcoinlib==0.5.0
cachetools==4.2.0
cashaddress-regtest==1.1.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cmake==3.17.2
coincurve==13.0.0
coinmarketcap==5.0.3
crypto==1.4.1
cryptography==3.2.1
cytoolz==0.11.0
decorator==4.4.2
defusedxml==0.7.0rc1
Deprecated==1.2.10
dj-database-url==0.5.0
dj-stripe==2.4.1
Django==3.1.4
django-ads==1.1.1
django-appconf==1.0.4
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cc==0.2.3
django-classy-tags==2.0.0
django-cors-headers==3.6.0
django-crispy-forms==1.10.0
django-embed-video==1.4.0
django-jquery==3.1.0
django-js-reverse==0.9.1
django-phone-login==1.1.3
django-phone-verify==2.0.0
django-phonenumber-field==1.1.0
django-qr-code==2.0.1
django-rest-framework==0.1.0
django-sekizai==2.0.0
django-sendsms==0.2.3
django-sslserver==0.22
Django-Verify-Email==0.0.4
djangorestframework==3.12.2
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.9.5
fastecdsa==2.1.5
firebase==3.0.1
Flask==1.1.2
future==0.18.2
gcloud==0.17.0
geocoder==1.38.1
google-auth==1.24.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
heroku==0.1.4
hexbytes==0.2.1
httplib2==0.18.1
idna==2.10
ipfshttpclient==0.7.0a1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsonfield==3.1.0
jsonschema==3.2.0
jwcrypto==0.8
jws==0.1.3
Kivy==1.11.1
kombu==5.0.2
lazy-object-proxy==1.5.0
lru-dict==1.1.6
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
multiaddr==0.0.9
Naked==0.1.31
netaddr==0.8.0
nexmo==2.5.2
oauth2client==3.0.0
oauthlib==3.1.0
parsimonious==0.8.1
parso==0.7.0
paypal==1.2.5
paystackapi==2.0.0
phonenumbers==8.12.15
phonenumberslite==8.12.15
Pillow==8.0.1
prompt-toolkit==3.0.8
protobuf==3.14.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoin==0.90.20201031
pycparser==2.20
pycryptodome==3.4.3
pygame==2.0.0.dev8
pygame-sdl2==2.1.0
pyjnius==1.2.1
PyJWT==1.7.1
pylint==2.5.2
Pyrebase==3.0.27
pyrsistent==0.17.3
PySDL2==0.9.7
python-bitcoinrpc==1.0
python-coincap==0.0.3
python-dateutil==1.5
python-dotenv==0.15.0
python-firebase==1.2
python-jwt==2.0.1
python3-openid==3.2.0
PythonVideoConverter==1.0.3
pytz==2020.4
PyYAML==5.3.1
ratelim==0.1.6
rave-python==1.2.7
requests==2.25.1
requests-cache==0.5.2
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rlp==2.0.1
rsa==4.6
scrypt==0.8.17
segno==1.3.1
sentry-sdk==0.19.4
shellescape==3.8.1
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
SQLAlchemy==1.3.20
sqlparse==0.4.1
sseclient==0.0.27
stripe==2.55.1
toml==0.10.1
toolz==0.11.1
twilio==6.50.1
typed-ast==1.4.1
urllib3==1.26.2
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.13.1
websockets==8.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
